- applying the maven formatter plugin - moving the repository tests to a repository package
