+++ phoenicis-apps/src/test/java/org/phoenicis/apps/repository/NullRepositoryTest.java
@@ -16,7 +16,7 @@
  * 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA.
  */
 
-package org.phoenicis.apps;
+package org.phoenicis.apps.repository;
 
 import org.junit.Test;
 import org.phoenicis.apps.repository.NullRepository;
