+++ phoenicis-javafx/src/main/java/org/phoenicis/javafx/views/common/ThemeManager.java
@@ -52,17 +52,20 @@ public class ThemeManager {
     }
 
     /**
-     * sets the path of the CSS file containing the default category icons from the repository
-     *
-     * @param defaultCategoryIconsCss path of the default category icons from the repository
+     * sets the path of the CSS file containing the default category icons from
+     * the repository
+     * 
+     * @param defaultCategoryIconsCss
+     *            path of the default category icons from the repository
      */
     public void setDefaultCategoryIconsCss(String defaultCategoryIconsCss) {
         this.defaultCategoryIconsCss = defaultCategoryIconsCss;
     }
 
     /**
-     * returns the path of the CSS file containing the default category icons from the repository
-     *
+     * returns the path of the CSS file containing the default category icons
+     * from the repository
+     * 
      * @return path of the default category icons from the repository
      */
     public String getDefaultCategoryIconsCss() {
@@ -70,17 +73,20 @@ public class ThemeManager {
     }
 
     /**
-     * sets the path of the CSS file containing the default engine icons from the repository
-     *
-     * @param defaultEngineIconsCss path of the default engine icons from the repository
+     * sets the path of the CSS file containing the default engine icons from
+     * the repository
+     * 
+     * @param defaultEngineIconsCss
+     *            path of the default engine icons from the repository
      */
     public void setDefaultEngineIconsCss(String defaultEngineIconsCss) {
         this.defaultEngineIconsCss = defaultEngineIconsCss;
     }
 
     /**
-     * returns the CSS file containing the default engine icons from the repository
-     *
+     * returns the CSS file containing the default engine icons from the
+     * repository
+     * 
      * @return default engine icons from the repository
      */
     public String getDefaultEngineIconsCss() {
