+++ phoenicis-javafx/src/main/java/org/phoenicis/javafx/views/mainwindow/settings/FileAssociationsPanel.java
@@ -4,7 +4,7 @@ import javafx.scene.layout.VBox;
 
 /**
  * This class represents the "File Associations" settings category
- *
+ * 
  * @author marc
  * @since 23.04.17
  */
